my-learnings/digital-sign/requirements.txt
blinker==1.8.2
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
click==8.1.7
cryptography==43.0.1
Flask==3.0.3
idna==3.8
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
pycparser==2.22
python-decouple==3.8
requests==2.32.3
urllib3==2.2.2
Werkzeug==3.0.4
